requirements.txt
absl-py==2.3.1
accelerate==1.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
av==14.2.0
babel==2.17.0
beautifulsoup4==4.13.4
bitsandbytes==0.46.0
bleach==6.2.0
blobfile==3.0.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
consisid_eva_clip==1.0.2
contourpy==1.3.1
cycler==0.12.1
Cython==3.1.4
datasets==3.3.2
debugpy==1.8.14
decorator==5.2.1
decord==0.6.0
deepspeed==0.16.6
defusedxml==0.7.1
diffusers==0.27.2
dill==0.3.8
DistVAE==0.0.0b5
easydict==1.13
einops==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
facexlib==0.3.0
fastapi==0.115.11
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.17.0
filterpy==1.4.5
fire==0.7.0
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2024.12.0
ftfy==6.3.1
gdown==5.2.0
GPUtil==1.4.0
gradio==5.20.1
gradio_client==1.7.2
groovy==0.1.2
grpcio==1.76.0
h11==0.14.0
hf-xet==1.1.8
hf_transfer==0.1.9
hjson==3.1.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
imhist==0.0.4
importlib_metadata==8.6.1
imWatermark==0.0.2
iniconfig==2.3.0
invisible-watermark==0.2.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.6
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.14
kaleido==0.2.1
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
lazy_loader==0.4
lightning-utilities==0.14.3
llvmlite==0.44.0
lpips==0.1.4
lxml==5.3.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.3
moviepy==2.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.4.4
multiprocess==0.70.16
narwhals==1.28.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.13.0
nltk==3.9.1
notebook==7.4.0
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.0
omegaconf==2.3.0
onnx==1.19.0
onnxruntime==1.22.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peft==0.17.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.7
plotly==6.0.0
pluggy==1.6.0
portalocker==3.1.1
positional-encodings==6.0.3
prettytable==3.16.0
proglog==0.1.12
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==4.25.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
pyfacer==0.0.5
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytorch-lightning==2.5.1.post0
pytz==2025.1
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.24.0
ruff==0.11.9
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.46.2
stringzilla==4.0.11
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==3.0.1
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.2.18
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.19.1
tomesd==0.1.3
tomli==2.2.1
tomlkit==0.13.2
torchinfo==1.8.0
torchmetrics==1.7.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.40.1
triton==3.2.0
typer==0.15.3
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.2
validators==0.35.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
xformers==0.0.29.post3
xfuser==0.4.3.post3
xxhash==3.6.0
yarl==1.20.0
yunchang==0.6.3.post1
zipp==3.21.0
